add semester summary function
