+++ package.json
@@ -47,8 +47,10 @@
     "file-saver": "^2.0.5",
     "firebase": "^8.8.0",
     "firebase-admin": "^9.12.0",
+    "fraction.js": "^4.3.7",
     "html-to-draftjs": "^1.5.0",
     "jose": "^5.2.3",
+    "pandas-js": "^0.2.4",
     "prop-types": "^15.7.2",
     "react": "^17.0.2",
     "react-app-polyfill": "^2.0.0",
